fix: Route, RouteGeometry 분리 시 불필요한 지연이 되는 구조를 개선하기위해 Route 엔티티로 통일
